+++ app/lforms.js
@@ -1,12 +1,13 @@
 // Use IIFE so that strict mode is not at the file level
 (function() {
   'use strict';
-  angular.module('lformsWidget', [
-    'ngAnimate',
-    'ui.bootstrap',
-    'smoothScroll',
-    'autocompleteLhcMod'
-    ])
+  var Def = require('autocomplete-lhc');
+  var widgetDeps = ['smoothScroll','autocompleteLhcMod'];
+  if (Def._tooltip)
+    widgetDeps = [Def._animate, Def._popover, Def._tooltip].concat(widgetDeps);
+  else
+    widgetDeps = ['ngAnimate', 'ui.bootstrap'].concat(widgetDeps);
+  angular.module('lformsWidget', widgetDeps)
     .config(['$animateProvider', function ($animateProvider) {
       $animateProvider.classNameFilter(/has-ng-animate/);
     }])
@@ -37,3 +38,7 @@
       };
     });
 })();
+
+// Define the top-level namespace object
+var LForms = {};
+module.exports = LForms;