+++ app/scripts/lib/js-class.js
@@ -7,7 +7,7 @@
   var initializing = false, fnTest = /xyz/.test(function(){xyz;}) ? /\b_super\b/ : /.*/;
 
   // The base Class implementation (does nothing)
-  this.Class = function(){};
+  var Class = function(){};
 
   // Create a new Class that inherits from this class
   Class.extend = function(prop) {
@@ -61,4 +61,6 @@
 
     return Class;
   };
+
+  module.exports = Class;
 })();