+++ webpack.config.js
@@ -31,6 +31,7 @@ function commonConfig() {
 
 let configs = [];
 let fhirVersions = Object.keys(require('./app/scripts/fhir/versions'));
+let versionedDist = 'lforms-'+require('./package.json').version;
 for (let version of fhirVersions) {
   let entryFile = './app/scripts/fhir/'+version+'/fhirRequire.js';
 
@@ -42,17 +43,77 @@ for (let version of fhirVersions) {
 
   let minConfig = commonConfig();
   minConfig.entry = entryFile;
-  let versionedDist = 'lforms-'+require('./bower.json').version;
   minConfig.output.filename = './dist/'+versionedDist+'/fhir/'+version+'/lformsFHIR.min.js',
   minConfig.mode = 'production';
   configs.push(minConfig);
 }
 
-// autocomplete-lhc and (eventually) other ES6 code
-let es6Config = commonConfig();
-es6Config.entry = './app/scripts/index.js';
-es6Config.output.filename = './app/generated/scripts/autocomplete-lhc.js';
-es6Config.mode = 'none';
-configs.push(es6Config);
+// LForms and dependencies
+// The Bower package needs a single, transpiled lforms.js file that does
+// not include other bower packages (angular, etc.)
+let bowerConfig = commonConfig();
+bowerConfig.entry = './app/scripts/bower-index.js';
+bowerConfig.output.filename = './bower-dist/lforms.js';
+bowerConfig.output.library = 'LForms';
+bowerConfig.devtool = 'source-map';
+bowerConfig.mode = 'none';
+configs.push(bowerConfig);
+
+// The browser-ready dist package needs all of the dependencies in a single file
+// (except for the versioned FHIR files).
+let lformsConfig = commonConfig();
+lformsConfig.entry = './app/scripts/index.js';
+lformsConfig.output.path = require('path').resolve(__dirname, 'dist/'+versionedDist);
+lformsConfig.output.filename = 'lforms.min.js';
+lformsConfig.output.library = 'LForms';
+lformsConfig.devtool = 'source-map';
+lformsConfig.mode = 'production';
+//lformsConfig.mode = 'none';
+// For angular-ui-bootstrap, we need to pick up and process the CSS imports
+const MiniCssExtractPlugin = require("mini-css-extract-plugin");
+lformsConfig.plugins = [
+  new MiniCssExtractPlugin({
+    filename: "styles/lforms.css"
+  })
+];
+lformsConfig.module.rules.push({
+  test: /\.css$/,
+//  include: /node_modules/,
+  use: [
+    {
+      // This loader creates one CSS file per JS file that contains CSS.
+      // Or so the documentation says.  It seems to just create one file.
+      loader: MiniCssExtractPlugin.loader,
+      options: {
+        outputPath: 'styles',
+      }
+    },
+    //'style-loader', // for devlopment only
+    'css-loader' // resolves paths for CSS files in require/import
+  ]
+});
+lformsConfig.module.rules.push({
+  test: /glyphicons.*\.(eot|svg|ttf|woff2?)$/,
+  use: [{
+    loader: 'file-loader',
+    options: {
+      name: '[name]_[hash].[ext]',
+      outputPath: 'styles/fonts',
+      publicPath: 'fonts'
+    }
+  }]
+});
+lformsConfig.module.rules.push({
+  test: /\.(png|svg|jpg|gif)$/,
+  use: [{
+    loader: 'file-loader',
+    options: {
+      name: '[name]_[hash].[ext]',
+      outputPath: 'styles/images',
+      publicPath: 'images'
+    }
+  }]
+});
+configs.push(lformsConfig);
 
 module.exports = configs;